+++ apps/client/src/composables/passports.spec.ts
@@ -107,13 +107,13 @@ describe("passports", () => {
     mocks.fetchPassports.mockResolvedValueOnce({ data: passportsResponse });
     await fetchPassports(
       { limit: 10, cursor: undefined },
-      { status: DigitalProductDocumentStatusDto.Archived },
+      { status: [DigitalProductDocumentStatusDto.Archived] },
     );
 
     expect(mocks.fetchPassports).toHaveBeenCalledWith({
       pagination: { limit: 10, cursor: undefined },
       populate: [Populates.assetAdministrationShells],
-      filter: { status: DigitalProductDocumentStatusDto.Archived },
+      filter: { status: [DigitalProductDocumentStatusDto.Archived] },
     });
     expect(passports.value).toEqual(passportsResponse);
   });