+++ apps/main/src/templates/presentation/template.controller.ts
@@ -661,7 +661,7 @@ export class TemplateController
     @LimitQueryParam() limit: number | undefined,
     @CursorQueryParam() cursor: string | undefined,
     @PopulateQueryParam() populate: string[],
-    @StatusQueryParam() status: DigitalProductDocumentStatusDtoType | undefined,
+    @StatusQueryParam() status: DigitalProductDocumentStatusDtoType[] | undefined,
     @OrganizationId() organizationId: string,
     @UserRoleDecorator() userRole: UserRoleType,
     @MemberRoleDecorator() memberRole: MemberRoleType | undefined,