+++ apps/client/CHANGELOG.md
@@ -1,5 +1,69 @@
 # @open-dpp/client
 
+## 0.4.0
+
+### Minor Changes
+
+- [#513](https://github.com/open-dpp/open-dpp/pull/513) [`ca55ba0`](https://github.com/open-dpp/open-dpp/commit/ca55ba0529752cdd852ba41fcc357dfc1b27bacb) Thanks [@florianBieck](https://github.com/florianBieck)! - Updated dependencies to minor and patch versions
+
+### Patch Changes
+
+- Updated dependencies [[`ca55ba0`](https://github.com/open-dpp/open-dpp/commit/ca55ba0529752cdd852ba41fcc357dfc1b27bacb)]:
+  - @open-dpp/api-client@0.4.0
+  - @open-dpp/dto@0.4.0
+
+## 0.3.0
+
+### Minor Changes
+
+- [#530](https://github.com/open-dpp/open-dpp/pull/530) [`d0959ed`](https://github.com/open-dpp/open-dpp/commit/d0959edc7a592bc46b83acfd397a011ae7412d08) Thanks [@mr42](https://github.com/mr42)! - Fix product image gallery in the presentation view such that all images regardless of their size fit the view.
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @open-dpp/api-client@0.3.0
+  - @open-dpp/dto@0.3.0
+
+## 0.2.2
+
+### Patch Changes
+
+- [#514](https://github.com/open-dpp/open-dpp/pull/514) [`5f8a7ff`](https://github.com/open-dpp/open-dpp/commit/5f8a7ff23a611237652e9bc9e01a5be97ef445d1) Thanks [@Hentra](https://github.com/Hentra)! - Add an option to set the branding color for an organization
+
+- Updated dependencies [[`5f8a7ff`](https://github.com/open-dpp/open-dpp/commit/5f8a7ff23a611237652e9bc9e01a5be97ef445d1)]:
+  - @open-dpp/api-client@0.2.2
+  - @open-dpp/dto@0.2.2
+
+## 0.2.1
+
+### Patch Changes
+
+- [#529](https://github.com/open-dpp/open-dpp/pull/529) [`daa1f4e`](https://github.com/open-dpp/open-dpp/commit/daa1f4e9ff07ce4e6d826c1a0490af4dcc3843e1) Thanks [@florianBieck](https://github.com/florianBieck)! - Fixed trusted publishing
+
+- Updated dependencies []:
+  - @open-dpp/api-client@0.2.1
+  - @open-dpp/dto@0.2.1
+
+## 0.2.0
+
+### Minor Changes
+
+- [#503](https://github.com/open-dpp/open-dpp/pull/503) [`40c772c`](https://github.com/open-dpp/open-dpp/commit/40c772c5015aff47c0e91c443860ae25941bd44f) Thanks [@mr42](https://github.com/mr42)! - Templates and passports have a status (Draft, Published, Archived) now.
+
+  The following operations are possible:
+
+  - publish (Draft -> Published)
+  - archive (Draft -> Archived or Published -> Archived)
+  - restore (Archived -> Published or Archived -> Draft)
+
+  In addition, a user can delete templates/ passports with the status draft.
+
+### Patch Changes
+
+- Updated dependencies [[`40c772c`](https://github.com/open-dpp/open-dpp/commit/40c772c5015aff47c0e91c443860ae25941bd44f)]:
+  - @open-dpp/api-client@0.2.0
+  - @open-dpp/dto@0.2.0
+
 ## 0.1.4
 
 ### Patch Changes