+++ packages/env/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @open-dpp/env
 
+## 0.4.0
+
+### Minor Changes
+
+- [#513](https://github.com/open-dpp/open-dpp/pull/513) [`ca55ba0`](https://github.com/open-dpp/open-dpp/commit/ca55ba0529752cdd852ba41fcc357dfc1b27bacb) Thanks [@florianBieck](https://github.com/florianBieck)! - Updated dependencies to minor and patch versions
+
+## 0.3.0
+
+## 0.2.2
+
+## 0.2.1
+
+## 0.2.0
+
 ## 0.1.4
 
 ### Patch Changes